Updated Scene 08
So the text for scene 08 was good but I was missing some slides to go with that so I added all of them here. There is still slides that are missing output but right now I'm delaying inserting that data until I have a live working demo and right now I'm working without a net.