Major Update Scene 04
A version control system discussion leads off the session. That discussion prompts the attendees to install git at that moment and not later when working in the cookbook.

There are steps now included for when git fails to work on the system and demands that git is setup on the system.

Addresses the issue opened in #22 of having the poorly named cookbooks.
Adds some information to the NOTES which resemble a training guide