Powerpoint Automation *BUGS*
So this is the initial implementation of opening and manipulating PowerPoint files.

These rake tasks will merge powerpoints from section 01 to the day_out outline presentation. It brings in all the slides but it does something wrong with the ordering because I believe that the Powerpoint data for the presentation.xml needs to have more consistency in the ordering as well.