FIX: Scene 01 Notes and Rakefile *BUGS*
So the first scene has a notes section and that is causing import issues because it is mentioned as a dependency and it is no longer present.

The second issue of wrongly ordering of the slides is fixed. A number of issues with being off by one and inserting the slides into the existing document. Now there are still errors with the sections. A re-ordering keeps happening.