Rakefile refactored to Module *BUGS*
Moved the Rakefile methods that I created to a module that will now merge the presentation together. This is still very functional programming but in the interest of lettin the concepts and classes make themselves known it feels like the right way to make the change.

I have not tested this for adding slides to a different section than the first section. That is going to happen in the next commit.