Refactored more into the Presentation
The first refactoring of the code I moved methods that were purely functional into the Presentation class. Now they still aren't right. This is the kind of data that I wouldn't really want to become dependent. But this is an initial step as these messages that the merger object is sending is still dependent on those results.