Fluet methods and PresentationMerger is a class
I wanted to have a syntax that was more of a fluent based approach. To emphasize the sending of messages. So the outline presentation can merge and presentation deck into this one at a particular section.

  outline_presentation.merge(presentation).into(section: 1)

So the merge method doesn't do all the work. It simply creates an instance of the PresentationMerger object and it still maintains all the methods now with a little bit of state (the two presentations which it uses in the #merge_into_section method).

There are still some methods that stand out as being functional. But they may wait for a moment longer as I figure out the structure of how I best to represent the interaction with messages.