Modeling the Files
So there is definitely some modeling of the files that needs to happen. so I started with these rudimentary classes. There is already a couple of patterns emerging with the managmenet of the files. That is something I will address later.

Here I continue to move more of the methods out of the Merger class and into the specific file that need to manage this content. There is a strangeness about caching that I have to be aware of because if I don't maintain instances I could lose all the xml data that does not get saved.

As I look at this code before sleeping I think this merger object should be able to send a message to the presentation object and ask for its underlying filesystem (probably not right in the long run). The filesystem object maintains references and ways to access all of the various files. This would also be a good place to move the methods for copying and shifting files.

The presentation merge should remain a coordinator, a controller with all the calls necessary to coordinate the primitives of the other classes. That will ensure the most flexibility.